+++ py/bitbox02/bitbox02/bitbox02/__init__.py
@@ -16,7 +16,7 @@
 from __future__ import print_function
 import sys
 
-__version__ = "6.0.0"
+__version__ = "6.1.0"
 
 if sys.version_info.major != 3 or sys.version_info.minor < 6:
     print(