+++ py/bitbox02/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [Unreleased]
+- Add `display` parameter to `root_fingerprint()`.
+
 ## 6.0.0
 - Offset the recoverable ID by 27 in the signature returned by `eth_sign_msg()`.
 - Rename `BTCOutputExternal.hash` to `BTCOutputExternal.payload`.